requirements.txt
asgiref==3.8.1
astroid==2.4.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Brotli==1.1.0
cairocffi==1.7.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.3
cssselect2==0.7.0
cycler==0.11.0
demjson3==3.0.6
Django==4.2.23
django-materializecss-form==1.1.17
django-sass-compiler==1.1.0
django-session-timeout==0.1.0
django-time-out==0.1.8
django-widget-tweaks==1.4.8
docopt==0.6.2
fonttools==4.37.1
h11==0.13.0
html5lib==1.1
idna==3.3
isort==5.5.0
jtutils==0.0.8
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
leven==1.0.4
libsass==0.20.1
matplotlib==3.5.3
mccabe==0.6.1
nose==1.3.7
num2words==0.5.14
numpy==1.23.2
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pillow==10.4.0
pycparser==2.21
pydyf==0.11.0
pylint==2.6.0
PyMySQL==1.1.2
pyparsing==3.0.9
pyphen==0.16.0
PySocks==1.7.1
python-csv==0.0.13
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2020.1
requests==2.28.1
selenium==4.4.3
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
tinycss2==1.4.0
toml==0.10.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.12
weasyprint==61.2
webencodings==0.5.1
wrapt==1.12.1
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.13.0
zopfli==0.2.3.post1
